Pin version of yarl
yarl==1.2.0 breaks aiohttp<=2.2.0 which is required by botocore==0.4.0. In the event that botocore==0.4.0 is a co-dependency of this library, pip will ignore the version constraint on yarl in aiohttp<=2.2.0 and install the latest version as originally specified here. The easiest way to fix this mismatch is pin the dependency here.